+++ src/main/java/com/part4/team09/otboo/module/domain/clothes/repository/custom/CustomClothesRepository.java
@@ -2,7 +2,9 @@ package com.part4.team09.otboo.module.domain.clothes.repository.custom;
 
 import com.part4.team09.otboo.module.common.enums.SortDirection;
 import com.part4.team09.otboo.module.domain.clothes.dto.data.ClothesWithAttributesDto;
+import com.part4.team09.otboo.module.domain.clothes.entity.Clothes;
 import com.part4.team09.otboo.module.domain.clothes.entity.Clothes.ClothesType;
+import com.part4.team09.otboo.module.domain.recommendation.dto.ClothingOption;
 import java.util.List;
 import java.util.UUID;
 
@@ -12,4 +14,6 @@ public interface CustomClothesRepository {
     ClothesType typeEqual, UUID ownerId, String sortBy, SortDirection sortDirection);
 
   List<ClothesWithAttributesDto> findByClothesId(UUID clothesId);
+
+  List<Clothes> findAllOrderedByAttributeScores(List<ClothingOption> options, int limit);
 }
