+++ src/main/java/com/part4/team09/otboo/module/domain/clothes/repository/custom/impl/CustomClothesRepositoryImpl.java
@@ -9,9 +9,14 @@ import com.part4.team09.otboo.module.domain.clothes.entity.QClothesAttribute;
 import com.part4.team09.otboo.module.domain.clothes.entity.QClothesAttributeDef;
 import com.part4.team09.otboo.module.domain.clothes.entity.QSelectableValue;
 import com.part4.team09.otboo.module.domain.clothes.repository.custom.CustomClothesRepository;
+import com.part4.team09.otboo.module.domain.recommendation.dto.ClothingOption;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.Projections;
+import com.querydsl.core.types.dsl.CaseBuilder;
+import com.querydsl.core.types.dsl.Expressions;
+import com.querydsl.core.types.dsl.NumberExpression;
+import com.querydsl.core.types.dsl.StringExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -133,6 +138,67 @@ public class CustomClothesRepositoryImpl implements CustomClothesRepository {
       .leftJoin(clothesAttributeDef).on(clothesAttributeDef.id.eq(selectableValue.attributeDefId))
       .where(clothes.id.eq(clothesId))
       .fetch();
+  }
+
+  @Override
+  public List<Clothes> findAllOrderedByAttributeScores(List<ClothingOption> options, int limit) {
+    QClothes c = QClothes.clothes;
+    QClothesAttribute ca = QClothesAttribute.clothesAttribute;
+    QSelectableValue sv = QSelectableValue.selectableValue;
+    QClothesAttributeDef cad = QClothesAttributeDef.clothesAttributeDef;
+
+    NumberExpression<Integer> totalScoreExpr = buildScoreExpression(options, cad.name, sv.item);
+
+    return queryFactory
+        .select(c)
+        .from(c)
+        .join(ca).on(ca.clothesId.eq(c.id))
+        .join(sv).on(sv.id.eq(ca.selectableValueId))
+        .join(cad).on(cad.id.eq(sv.attributeDefId))
+        .groupBy(c.id)
+        .orderBy(totalScoreExpr.asc())
+        .limit(limit)
+        .fetch();
+  }
+
+  private NumberExpression<Integer> buildScoreExpression(
+      List<ClothingOption> options,
+      StringExpression attributeName,
+      StringExpression itemValue) {
+    CaseBuilder caseBuilder = new CaseBuilder();
+    CaseBuilder.Cases<Integer, NumberExpression<Integer>> cases = null;
+
+    for (ClothingOption option : options) {
+      String attr = option.attribute();
+      List<String> values = option.values();
+
+      for (int i = 0; i < values.size(); i++) {
+        String value = values.get(i);
+
+        NumberExpression<Integer> scoreExpr =
+            Expressions.numberTemplate(Integer.class, "{0}", i + 1);
+
+        if (cases == null) {
+          cases = caseBuilder
+              .when(attributeName.eq(attr).and(itemValue.eq(value)))
+              .then(scoreExpr);
+        } else {
+          cases = cases
+              .when(attributeName.eq(attr).and(itemValue.eq(value)))
+              .then(scoreExpr);
+        }
+      }
+    }
+
+    if (cases == null) {
+      return Expressions.numberTemplate(Integer.class, "{0}", 999);
+    }
+
+    return Expressions.numberTemplate(
+        Integer.class,
+        "SUM({0})",
+        cases.otherwise(999)
+    );
   }
 
   private OrderSpecifier<?> getOrderSpecifier(String sortBy, SortDirection sortDirection) {
